Added option to pass a rdd to workloads. Added blocking workload with reading
Makes it possible to perform different workloads on a rdd. This way it's
possible to create a "pipeline" of workloads on one rdd which is closer
to a real workload.
The new workload "distributed-blocking-with-reading" makes it possible
to block tasks depending on a distribution and loading data before it.
This adds the complexity of data transfers to the blocking workload.
Depending on the data structure this complexity can be included by
changing the parameters of the distribution.